test/requirements.txt
pandas
yfinance
streamlit
ccxt
python-binance